refactor(metrics): decouple Portal recording from exporters
